frontend/python/requirements.txt
fastapi
uvicorn
python-dotenv
autogen
twilio
pydantic